IT WORKS!!!!!!!!!
The game is feature complete! All that needs done now is to figure out how to keep track of credits and fix a small bug with two pair hands where if either pair is face cards it shows that it's a single pair of jacks or better (which, while technically true, is inaccurate).